Move ChannelDefinitionProcessor to API package
