+++ modules/flowable-event-registry/src/main/java/org/flowable/eventregistry/impl/DefaultEventRegistry.java
@@ -20,6 +20,7 @@ import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 import org.flowable.common.engine.api.FlowableIllegalArgumentException;
 import org.flowable.common.engine.impl.interceptor.EngineConfigurationConstants;
+import org.flowable.eventregistry.api.ChannelDefinitionProcessor;
 import org.flowable.eventregistry.api.CorrelationKeyGenerator;
 import org.flowable.eventregistry.api.EventRegistry;
 import org.flowable.eventregistry.api.EventRegistryEvent;
@@ -31,7 +32,6 @@ import org.flowable.eventregistry.api.OutboundEventProcessor;
 import org.flowable.eventregistry.api.model.InboundChannelDefinitionBuilder;
 import org.flowable.eventregistry.api.model.OutboundChannelDefinitionBuilder;
 import org.flowable.eventregistry.api.runtime.EventInstance;
-import org.flowable.eventregistry.impl.deployer.ChannelDefinitionProcessor;
 import org.flowable.eventregistry.impl.model.InboundChannelDefinitionBuilderImpl;
 import org.flowable.eventregistry.impl.model.OutboundChannelDefinitionBuilderImpl;
 import org.flowable.eventregistry.model.ChannelDefinition;
